Change int64 to long long to avoid errors on Mac

+++ src/scheduler_callback_thread.h
@@ -6,13 +6,13 @@
 #include "../JuceLibraryCode/JuceHeader.h"
 
 // FIXME: this should go into a header file
-extern int64 sp_midi_get_current_time_microseconds();
+extern long long sp_midi_get_current_time_microseconds();
 
 
 // FIXME: make sure this is thread-safe (i.e. protect the m_pending_messages accesses)
 class SchedulerCallbackThread : public Thread
 {
-    using TupleTimePidInteger = std::tuple<int64, ErlNifPid, int64>;
+    using TupleTimePidInteger = std::tuple<long long, ErlNifPid, long long>;
 public:
     SchedulerCallbackThread() : Thread("scheduler callback thread") {
         setPriority(10);
@@ -20,7 +20,7 @@ public:
 
     void run() override
     {
-        int64 next_timer_shot = 0;
+        long long next_timer_shot = 0;
         while (!threadShouldExit()){
             // If there are no pending messages, do nothing
             bool is_empty;
@@ -39,8 +39,8 @@ public:
             next_timer_shot = first_pending_message->get_time();
 
             // How far is it in the future?
-            int64 current_time = sp_midi_get_current_time_microseconds();
-            int64 time_to_next_message_millis = (next_timer_shot - current_time) / 1000;
+            long long current_time = sp_midi_get_current_time_microseconds();
+            long long time_to_next_message_millis = (next_timer_shot - current_time) / 1000;
 
             // If it is less than 1ms away or in the past, fire it
             if (time_to_next_message_millis < 1) {
@@ -58,11 +58,11 @@ public:
             }
             
             // Wait for 10 ms if there is nothing happening soon, or as necessary if there is something less than 10ms away
-            wait(jmin(static_cast<int64>(10), time_to_next_message_millis));
+            wait(jmin(static_cast<long long>(10), time_to_next_message_millis));
         }
     }
 
-    void trigger_callback_at(int64 when, ErlNifPid pid, int64 integer)
+    void trigger_callback_at(long long when, ErlNifPid pid, long long integer)
     {   
         TupleTimePidIntegerWrapper element(when, pid, integer);
         std::lock_guard<std::mutex> guard(m_messages_mutex);  
@@ -73,16 +73,16 @@ private:
 
     class TupleTimePidIntegerWrapper {
     public:
-        TupleTimePidIntegerWrapper(int64 when, ErlNifPid pid, int64 integer) : m_when(when), m_pid(pid), m_integer(integer) { };
-        int64 get_time() const { return m_when; };
+        TupleTimePidIntegerWrapper(long long when, ErlNifPid pid, long long integer) : m_when(when), m_pid(pid), m_integer(integer) { };
+        long long get_time() const { return m_when; };
         ErlNifPid get_pid() const { return m_pid; };
-        int64 get_integer() const { return m_integer; };
+        long long get_integer() const { return m_integer; };
 
     private:
         TupleTimePidInteger tuple;
-        int64 m_when;
+        long long m_when;
         ErlNifPid m_pid;
-        int64 m_integer;
+        long long m_integer;
 
         friend bool operator<(const TupleTimePidIntegerWrapper& l, const TupleTimePidIntegerWrapper& r) {
             return l.m_when < r.m_when;
@@ -92,7 +92,7 @@ private:
     std::set<TupleTimePidIntegerWrapper> m_pending_messages;
     std::mutex m_messages_mutex;
 
-    int send_integer_to_erlang_process(ErlNifPid pid, int64 integer, int64 time)
+    int send_integer_to_erlang_process(ErlNifPid pid, long long integer, long long time)
     {
         ErlNifEnv *msg_env = enif_alloc_env();
         ERL_NIF_TERM term, term2, list;
